refactor: mission create page
